requirements.txt
aiomysql==0.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
fastapi-jwt-auth==0.5.0
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
fastapi-versioning==0.10.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
Jinja2==3.1.2
makefun==1.15.1
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
passlib==1.7.4
protobuf==3.20.3
psycopg==3.1.9
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.5.4
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
